requirements.txt
asgiref==3.8.1
Django==4.2
django-admin-sortable2==2.1.10
django-dotenv==1.4.2
django-filebrowser==4.0.3
django-grappelli==3.0.8
django-jalali==6.0.1
django-solo==2.2.0
django-taggit==5.0.1
django-tinymce==3.7.1
django-webpack-loader==3.0.1
jdatetime==4.1.1
pillow==10.2.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
